macos: delete Ghostty.SplitView
